Revert "Writable nested relations implemented for student endpoint"
