+++ apps/schools/api_views/institution.py
@@ -2,44 +2,33 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import generics, viewsets
 
-from schools.filters import (
-    BoundaryFilter, ProgrammeFilter, AssessmentFilter,
-    InstitutionFilter,
-)
+from schools.api_filters import BoundaryFilter
 from schools.serializers import (
     InstitutionSerializer, AssessmentSerializer, ProgrammeSerializer,
-    BoundarySerializer, StaffSerializer
+    BoundarySerializer
 )
 from schools.models import (
     Institution, AssessmentInstitution, ProgrammeInstitution,
-    Boundary, Staff
+    Boundary
 )
 
 
 class InstitutionViewSet(viewsets.ModelViewSet):
     queryset = Institution.objects.all()
     serializer_class = InstitutionSerializer
-    filter_class = InstitutionFilter
 
 
 class AssessmentViewSet(viewsets.ModelViewSet):
     queryset = AssessmentInstitution.objects.all()
     serializer_class = AssessmentSerializer
-    filter_class = AssessmentFilter
 
 
 class ProgrammeViewSet(viewsets.ModelViewSet):
     queryset = ProgrammeInstitution.objects.all()
     serializer_class = ProgrammeSerializer
-    filter_class = ProgrammeFilter
 
 
 class BoundaryViewSet(viewsets.ModelViewSet):
     queryset = Boundary.objects.all()
     serializer_class = BoundarySerializer
     filter_class = BoundaryFilter
-
-class StaffViewSet(viewsets.ModelViewSet):
-    queryset = Staff.objects.all()
-    serializer_class=StaffSerializer
-
