--- apps/tests/__init__.py
@@ -1,1 +1,0 @@
-from .test_students import StudentsApiTestCase
