+++ tada/api_urls.py
@@ -2,7 +2,8 @@ from rest_framework import routers
 
 from django.conf.urls import patterns, url
 
-from schools.api_views import (StaffViewSet,InstitutionViewSet, StudentViewSet,
+from schools.api_views import (
+    InstitutionViewSet, StudentsListViewSet,
     AssessmentViewSet, ProgrammeViewSet, BoundaryViewSet
 )
 
@@ -12,6 +13,5 @@ router.register(r'institutions', InstitutionViewSet, base_name='institution')
 router.register(r'assessments', AssessmentViewSet, base_name='assessment')
 router.register(r'programmes', ProgrammeViewSet, base_name='programme')
 router.register(r'boundaries', BoundaryViewSet, base_name='boundary')
-router.register(r'staff',StaffViewSet, base_name='staff')
-router.register(r'students', StudentViewSet, base_name='students')
+router.register(r'students', StudentsListViewSet, base_name='students')
 urlpatterns = router.urls
